issue-539: extract writedata actor into separate location + add TABLET_VERIFY for ReleaseCollectBarrier
