+++ cloud/filestore/libs/storage/tablet/actors/ya.make
@@ -1,0 +1,12 @@
+LIBRARY()
+
+
+SRCS(
+    tablet_writedata.cpp
+)
+
+PEERDIR(
+    cloud/filestore/libs/storage/tablet/model
+)
+
+END()
