+++ cloud/filestore/libs/storage/tablet/tablet_state.h
@@ -765,7 +765,7 @@ public:
     }
 
     void AcquireCollectBarrier(ui64 commitId);
-    void ReleaseCollectBarrier(ui64 commitId);
+    bool TryReleaseCollectBarrier(ui64 commitId);
 
     ui64 GetCollectCommitId() const;
